update appEligibilityInfo interface
