+++ packages/teams-js/src/private/copilot.ts
@@ -1,12 +1,12 @@
 import { ensureInitialized } from '../internal/internalAPIs';
 import { errorNotSupportedOnPlatform } from '../public/constants';
 import { runtime } from '../public/runtime';
-import { AppEligibilityInformation } from './interfaces';
+import { AppEligibilityInformation } from '../public/interfaces';
 
 /**
  * @beta
  * @hidden
- * Namespace to delegate M365 chat app specific APIs
+ * Namespace to delegate copilot app specific APIs
  * @internal
  * Limited to Microsoft-internal use
  */
@@ -24,8 +24,7 @@ export namespace copilot {
      * @internal
      * Limited to Microsoft-internal use
      * @beta
-     * This function is called by M365Chat app
-     * @returns boolean to represent whether externalAppCommands capability is supported
+     * @returns boolean to represent whether copilot.eligibility capability is supported
      *
      * @throws Error if {@linkcode app.initialize} has not successfully completed
      */
@@ -38,8 +37,7 @@ export namespace copilot {
      * @internal
      * Limited to Microsoft-internal use
      * @beta
-     * This function is called by M365Chat app
-     * @returns the M365Chat eligibility information about the user
+     * @returns the copilot eligibility information about the user
      *
      * @throws Error if {@linkcode app.initialize} has not successfully completed
      */
